outlook-email-mcp/requirements.txt
mcp>=1.2.0
pywin32>=228
requests>=2.28.0
python-dotenv>=0.21.0
